requirements-backend.txt
# Backend API Requirements
# FastAPI and server
fastapi>=0.110
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0  # PostgreSQL driver (v3) - use postgresql+psycopg:// in connection string
alembic>=1.13.0
pgvector>=0.2.4

# Data validation and settings
pydantic>=2.0.0
pydantic-settings>=2.1.0
pydantic[email]>=2.0.0

# Authentication
PyJWT>=2.8.0

# HTTP client
httpx>=0.27
requests>=2.31.0

# Environment and config
python-dotenv>=1.0.0

# Supabase (if needed)
supabase>=2.4

# Testing
pytest>=7.4.0

# Multipart form support
python-multipart>=0.0.6

